Div updates and improvements
The bubbles starts to animate as soon as they are created even if they are created one by one. It makes a lot more sense when I want to make graph showing the decay in fps as more bubbles are added and animated.
Fixed some code.  Removed some old code I'm not going to use. Still struggling a bit getting Bootstrap to be completely fluid. There is a but in the transformation from horizontal to vertical.